Add asynchronous query functionality with XBRLAsync class and refactor existing query method
