+++ src/xbrl_us/experimental/__init__.py
@@ -1,0 +1,3 @@
+from .aquery import XBRLAsync
+
+__all__ = ["XBRLAsync"]
